requirements.txt
astroid==2.4.2
blinker==1.4
docutils==0.16
feedgenerator==1.9.1
isort==5.7.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
pelican==4.5.4
Pygments==2.6.1
pylint==2.6.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.5
six==1.15.0
text-unidecode==1.3
toml==0.10.2
Unidecode==1.2.0
wrapt==1.12.1
